fixed color for memory using type casting

+++ src/utils.c
@@ -125,8 +125,8 @@ void update_texture(SDL_Texture* texture) {
     uint16_t val = memory[addr];
     uint16_t intensity = (uint16_t)(val >> BIT_SHIFT_8) & BYTE_MASK;
     uint32_t color = (BYTE_MASK << BIT_SHIFT_24) |            // A
-                     (uint16_t)(intensity << BIT_SHIFT_16) |  // R
-                     (uint16_t)(intensity << BIT_SHIFT_8) |   // G
+                     (uint32_t)(intensity << BIT_SHIFT_16) |  // R
+                     (uint32_t)(intensity << BIT_SHIFT_8) |   // G
                      (intensity);                             // B
 
     pixel_ptr[addr] = color;
